+++ Sources/KeyboardKit/Callouts/Callouts.swift
@@ -11,7 +11,8 @@ import Foundation
 /**
  This namespace contains callout-related types.
  
- The namespace doesn't contain protocol, observable types or
- open classes, not types that are meant to be top-level ones.
+ The namespace doesn't contain protocols or open classes, or
+ types that are meant to be top-level ones. It's meant to be
+ a container for types used by top-level types.
  */
 public struct Callouts {}