+++ Sources/KeyboardKit/Emojis/Emojis.swift
@@ -11,7 +11,11 @@ import Foundation
 /**
  This namespace contains emojis-related types.
  
- The namespace doesn't contain protocol, observable types or
- open classes, not types that are meant to be top-level ones.
+ The namespace doesn't contain protocols or open classes, or
+ types that are meant to be top-level ones. It's meant to be
+ a container for types used by top-level types.
+ 
+ For now, this namespace will only have things in it when it
+ is part of the KeyboardKit Pro build.
  */
 public struct Emojis {}